+++ src/array.js
@@ -52,4 +52,9 @@ export const arrayToObject = _.curry((k, v, a) =>
   _.flow(_.keyBy(k), _.mapValues(v))(a)
 )
 
-export const flags = x => _.zipObject(x, _.times(x => true, x.length))
+// zipObject that supports functions instead of objects
+export const zipObjectDeepWith = _.curry((x, y) =>
+  _.zipObjectDeep(x, _.isFunction(y) && _.isArray(x) ? _.times(y, x.length) : y)
+)
+
+export const flags = zipObjectDeepWith(_, () => true)